allItems in UI by name
