+++ public/react/components/ItemList.js
@@ -1,10 +1,28 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { Item } from "./Item";
+import apiURL from "../api";
 
 export function ItemList() {
+    const [items, setItems] = useState([]);
+
+    async function fetchAllItems() {
+        try {
+            const response = await fetch(`${apiURL}/items`);
+            const itemsData = await response.json();
+            setItems(itemsData);
+        } catch (err) {
+            console.log('Error', err);
+        }
+    };
+
+    useEffect(() => {
+        fetchAllItems();
+    }, []);
     return (
         <>
-            <h1>ItemList</h1>
+            {items.map((item, idx) => (
+                <Item item={item} />
+            ))}
         </>
     )
 }
